Adequação do projeto para o modelo MVP.
