+++ app/src/main/AndroidManifest.xml
@@ -12,15 +12,15 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
 
-        <activity android:name=".SplashScreenActivity" >
+        <activity android:name=".View.SplashScreenActivity" >
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".MainActivity"/>
-        <activity android:name=".BeerActivity" />
+        <activity android:name=".View.MainActivity"/>
+        <activity android:name=".View.BeerActivity" />
     </application>
 
 </manifest>
