--- app/src/main/java/br/com/breno/beermenu/Interface/MyResult.kt
@@ -1,6 +1,0 @@
-package br.com.breno.beermenu.Interface
-
-/**Interface responsável por ligar os eventos do adapter com a mainActivity**/
-interface MyResult {
-    fun myResult(result: Int, message: String)
-}
