+++ app/src/main/res/layout/activity_main.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context=".MainActivity">
+    tools:context=".View.MainActivity">
 
     <android.support.v7.widget.Toolbar
         android:id="@+id/toolbarMain"
