+++ app/src/main/res/layout/activity_splash_screen.xml
@@ -5,7 +5,7 @@
     android:id="@+id/activity_splash"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".SplashScreenActivity">
+    tools:context=".View.SplashScreenActivity">
     <ImageView
         android:id="@+id/logo"
         app:srcCompat="@drawable/beer"
